Test for SSE, eslint fix and new routes
