+++ src/components/user/client.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {useState, useEffect} from 'react';
-import {Routes, Route, useNavigate} from 'react-router-dom';
 
 
 export default function Client({token, user, role}) {
