+++ src/models/users.js
@@ -4,7 +4,6 @@ const baseURL = "http://localhost:3002";
 
 const userModel = {
     login: async function login(user) {
-        console.log(user)
         try {
             const response = await fetch(`${baseURL}/v1/user/signin`, {
                 method: "POST",
@@ -13,7 +12,7 @@ const userModel = {
                     'content-type': 'application/json'
                 },
             });
-    
+            console.log(response)
             const result = await response.json();
             
             return result;
@@ -39,7 +38,7 @@ const userModel = {
         }
     },
     getAllCustomers: async function getAllCustomers() {
-        const result = fetch(`${baseURL}/v1/user/allUsers`)
+        const result = fetch(`${baseURL}/v1/user/all`)
             .then(r => r.json())
             .then(result => {return result})
             .catch((error) => {
