+++ src/routes/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import {useState, useEffect} from 'react';
-import {Routes, Route, useNavigate} from 'react-router-dom';
+import {useNavigate} from 'react-router-dom';
 
 import image1 from "../components/images/home.png";
 
@@ -15,7 +14,7 @@ export default function Home() {
     <div className='body'>
       <div className='kit'>
         <div className='homepageimages'>
-        <img src={image1} alt="image" className="homepicone"></img>
+        <img src={image1} alt="startimage" className="homepicone"></img>
         <button onClick={() => navigatereg()} className='buttonnewclient'>Registrera dig? Klicka här!</button>
         </div>
       </div>
