+++ src/routes/user.js
@@ -1,23 +1,17 @@
 import React from 'react';
 import {useState, useEffect} from 'react';
-import {Routes, Route, useNavigate} from 'react-router-dom';
 
 import Client from '../components/user/client.js';
 import AdminIndex from './admin/adminindex.js';
 
 
 export default function User({token, user, role}) {
-    const [ extraSection, setExtraSection ] = useState(null);
     const [comp, setComponent] = useState("")
-    const navigate = useNavigate();
 
-    //Update and set documents as initial stage.
     useEffect(() => {
         clickhandler("userdata")
-        addSection(role)
     }, []);
 
-    //To render one-page
     async function clickhandler(value) {
         if (value === "userdata"){
             setComponent("userdata")
@@ -28,34 +22,14 @@ export default function User({token, user, role}) {
         }
     };
 
-    async function addSection(value) {
-        if (value === "admin") {
-            setExtraSection("admin");
-        } else if (value === "customer") {
-            setExtraSection("customer");
-        }
-    };
-
-    function setCity(city){
-        if (city === "Visby"){
-            navigate('stad/Visby');
-        } else if (city === "Borlänge"){
-            navigate('stad/Borlänge');
-        } else if (city === "Lund"){
-            navigate('stad/Lund');
-        } else if (city === "Kund"){
-            navigate('kund');
-        }
-    }
-
     return (
         <div className='body'>
             <div className='kit'>
                 {token !== "" ?
                     <div>
-                    {extraSection === "admin" ?
+                    {role === "admin" ?
                         <AdminIndex/>
-                    : extraSection === "customer" ? 
+                    : role === "customer" ? 
                         <Client role={role} token={token} user={user}/>
                         : 
                         <div className='body'>
